feat(uffs-mft): capture $UsnJrnl:$J via sparse-compact run reader
$UsnJrnl:$J is a sparse file whose logical size can reach ~150 GB while
the live journal is a tiny allocated tail (the purged prefix is a
deallocated hole). The full-span reader hit the "too large to
materialize" guard and skipped it. Add read_runs_sparse_compact, which
reads only the non-sparse runs concatenated — each USN record is
self-describing, so dropping the leading hole loses no data. Share
run-resolution (base record + $ATTRIBUTE_LIST extensions) via
resolve_data_runs and the fallible allocation via alloc_zeroed.

Split the live-volume readers out of metafile.rs into a new
platform/metafile_read.rs (metafile.rs kept the on-disk format + save/
load); the growing reader set pushed metafile.rs to the 800-line gate.
read_metafile is re-exported so callers are unaffected.
